// src/payload-types.ts
export interface Post {
  id: number;
  title: string;
  slug: string;
  content?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  published_date?: string | null;
  tags?:
    | (
        | 'VPN Guide'
        | 'Proxy Tutorial'
        | 'Unblock Websites'
        | 'Privacy Tips'
        | 'Streaming Access'
        | 'Gaming Unblock'
        | 'Social Media Access'
        | 'Regional Restrictions'
        | 'Security Tips'
        | 'Tool Reviews'
      )[]
    | null;
  meta_title?: string | null;
  meta_description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}